chore: move splitListenAddr into server package
[upstream commit: 2e442eb88e1ec8b5ca30ca9fa99ab04bdc599d46]
